// UTest/pages/program-administration/models/review-document-list.model.ts
import { RvwAttachType } from "./rvw-attach-type.model";
import { RvwDocSubmissionReminderSchedule } from "./rvw-doc-submission-reminder-schedule.model";

export interface ReviewDocumentList {
  docListId: number;
  recType?: string;
  offCd?: string;
  grantPgmId?: number;
  seqNo: number;
  subRvwId?: number;
  docCat?: string;
  docCd?: string;
  docName?: string;
  docType?: string;
  valInd?: string;
  valDt?: Date;
  valUnits?: number;
  valUm?: string;
  minNos?: number;
  maxNos?: number;
  fileTypeAppl?: string;
  helpTxt?: string;
  fileTitle?: string;
  attachName?: string;
  createDt?: Date;
  createId?: number;
  lastUpdDt?: Date;
  lastUpdId?: number;
  subCat?: string;
  rvwType?: string;
  entType?: string;
  formId?: number;
  requireDocumentComment?: boolean;
  requireCheckListComment?: boolean;
  inactive?: boolean;
  defaultDoc?: string;
  documentDueDateDaysOffsetFromReviewStartDate?: number;
  flexFormPdftemplateId?: number;
  allowMultipleFlexformInstances?: boolean;
  unlimitedRecipients?: boolean;
  dataSrc?: string;
  approvalEmailTemplateId?: number;
  editableEmail?: boolean;
  approvalHtml?: string;
  approvalConfirmationEmailTemplateId?: number;
  questionnaireId?: number;
  allowMultipleQuestionnaireInstances?: boolean;
  questionnaireFieldId?: number;
  mapQuestionnaireWorkflowCriteria?: boolean;
  remapQuestionnaire?: boolean;
  rvwDocSubmissionReminderSchedules: RvwDocSubmissionReminderSchedule[];
  rvwAttachTypes: RvwAttachType[];
}